// language-platform-ui/src/app/model/payloadTokenResponse.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { TokenResponse } from "./models";


export interface PayloadTokenResponse { 
    data?: TokenResponse;
    error?: any;
    message?: string;
    status?: number;
    timestamp?: string;
}